Add fast, but potentially incorrect --fast flag to build.sh. Use with caution and never for distribution

+++ build-support/build.sh
@@ -4,11 +4,15 @@ scriptDir=$(cd $(dirname $0); pwd)
 rootDir=$scriptDir/..
 libDir=$rootDir/lib
 
+fast=""
+if (( $# == 1 )); then
+    fast=$1
+fi
+
+ZINC_OPTS="-scala-home $SCALA_HOME"
+
 # This line will also make sure SCALA_HOME is defined thanks to set -u
 echo >&2 "Using Scala install at $SCALA_HOME (Note: 2.10+ is required)"
-
-# Pass extra arguments to zinc
-ZINC_OPTS="-scala-home $SCALA_HOME $@"
 
 # Include both webui and test JARs
 libs=$(find $rootDir/lib/ -iname '*.jar')
@@ -29,13 +33,18 @@ find $rootDir/src/main/scala $rootDir/src/test/scala -iname '*.scala' \
 echo >&2 "Building initial JAR..."
 (cd $rootDir/bin/zinc; zip -qr $rootDir/ducttape.tmp.jar *)
 
-# Reads $rootDir/ducttape.tmp.jar as an intermediate input and writes $rootDir/ducttape.jar
 jarPath=$rootDir/ducttape.jar
-echo >&2 "Minimizing JAR using proguard..."
-time java -jar $scriptDir/proguard-4.8.jar @$scriptDir/proguard.ini
-
-# Add version.info
-(cd $rootDir; zip -gqr $jarPath version.info)
-rm $rootDir/ducttape.tmp.jar
-
+if [[ "$fast" == "--fast" ]]; then
+    echo >&2 "WARNING: Using fast (potentially incorrect) JAR builder; if you encounter errors, build again without --fast"
+    cd $rootDir/bin/zinc
+    zip -qu $jarPath *
+else
+    echo >&2 "Minimizing JAR using proguard..."
+    # Reads $rootDir/ducttape.tmp.jar as an intermediate input and writes $rootDir/ducttape.jar
+    time java -jar $scriptDir/proguard-4.8.jar @$scriptDir/proguard.ini
+    
+    # Add version.info
+    (cd $rootDir; zip -gqr $jarPath version.info)
+    rm $rootDir/ducttape.tmp.jar
+fi
 echo >&2 "Wrote JAR: $jarPath"
